pyproject.toml
[tool.poetry]
name = "streamlit-chatgpt"
version = "0.1.0"
description = ""
authors = ["h-k-nyosu <kuro.h1020@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.2"
python-dotenv = "^1.0.0"
streamlit = "^1.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
